scripts/adosync/requirements.txt
executor==22.0
python-dotenv==0.13

